change to MVP architecture pattern
